+++ logitacker/logitacker_options.h
@@ -31,6 +31,7 @@ typedef enum {
 typedef enum {
     OPTION_LOGITACKER_WORKMODE_UNIFYING,   // compatible to Unifying and (encrypted) presentation clickers
     OPTION_LOGITACKER_WORKMODE_LIGHTSPEED,   // compatible to Logitech LIGHTSPEED
+    OPTION_LOGITACKER_WORKMODE_G700,   // compatible to Logitech LIGHTSPEED
 } option_logitacker_workmode_t;
 
 typedef enum {