+++ logitacker/logitacker_processor_pair_device.c
@@ -177,6 +177,17 @@ void processor_pair_device_init_func_(logitacker_processor_pair_device_ctx_t *se
     nrf_esb_set_retransmit_delay(250);
     nrf_esb_set_tx_power(NRF_ESB_TX_POWER_8DBM);
 
+    // use special channel lookup table for pairing mode
+    switch (g_logitacker_global_config.workmode) {
+        case OPTION_LOGITACKER_WORKMODE_LIGHTSPEED:
+            nrf_esb_update_channel_frequency_table_lightspeed_pairing();
+            break;
+        case OPTION_LOGITACKER_WORKMODE_G700:
+        case OPTION_LOGITACKER_WORKMODE_UNIFYING:
+            nrf_esb_update_channel_frequency_table_unifying_pairing();
+            break;
+    }
+
 
     processor_pair_device_update_tx_payload_and_transmit(self);
 
@@ -317,6 +328,11 @@ void processor_pair_device_esb_handler_func_(logitacker_processor_pair_device_ct
                     p_dongle->is_texas_instruments = true;
                     p_dongle->classification = DONGLE_CLASSIFICATION_IS_LOGITECH_LIGHTSPEED;
                 }
+                if (p_dongle->wpid[0] == 0x80 && p_dongle->wpid[1] == 0x06) {
+                    p_dongle->is_nordic = true;
+                    p_dongle->classification = DONGLE_CLASSIFICATION_IS_LOGITECH_G700;
+                }
+
 
             }
             p_device->key_known = pi.key_material_complete;
@@ -357,20 +373,37 @@ void processor_pair_device_create_req1_pay(logitacker_processor_pair_device_ctx_
     memcpy(&self->tmp_tx_payload.data[3], pseudo_device_address, 5); // 3..7 pseudo device's current RF address
     self->tmp_tx_payload.data[8] = 0x08; //likely default keep-alive interval
     memcpy(&self->tmp_tx_payload.data[9], self->device_pairing_info.device_wpid, 2); //WPID of device
-    if (g_logitacker_global_config.workmode == OPTION_LOGITACKER_WORKMODE_UNIFYING) {
-        self->tmp_tx_payload.data[11] = LOGITACKER_DEVICE_PROTOCOL_UNIFYING; //likely protocol (0x04 == Unifying ?)
-        self->tmp_tx_payload.data[12] = 0x00; //unknown 0x00 for some devices (mouse, keyboard), 0x02 for others (mouse Anywhere MX 2, presenter)
-    } else {
-        self->tmp_tx_payload.data[11] = LOGITACKER_DEVICE_PROTOCOL_LIGHTSPEED;
-        self->tmp_tx_payload.data[12] = 0x02; //unknown 0x00 for some devices (mouse, keyboard), 0x02 for others (mouse Anywhere MX 2, presenter)
+    switch (g_logitacker_global_config.workmode) {
+        case OPTION_LOGITACKER_WORKMODE_UNIFYING:
+            self->tmp_tx_payload.data[11] = LOGITACKER_DEVICE_PROTOCOL_UNIFYING; //likely protocol (0x04 == Unifying ?)
+            self->tmp_tx_payload.data[12] = 0x00; //unknown 0x00 for some devices (mouse, keyboard), 0x02 for others (mouse Anywhere MX 2, presenter)
+            break;
+        case OPTION_LOGITACKER_WORKMODE_LIGHTSPEED:
+            self->tmp_tx_payload.data[11] = LOGITACKER_DEVICE_PROTOCOL_LIGHTSPEED;
+            self->tmp_tx_payload.data[12] = 0x02; //unknown 0x00 for some devices (mouse, keyboard), 0x02 for others (mouse Anywhere MX 2, presenter)
+            break;
+        case OPTION_LOGITACKER_WORKMODE_G700:
+            self->tmp_tx_payload.data[11] = LOGITACKER_DEVICE_PROTOCOL_G700;
+            self->tmp_tx_payload.data[12] = 0x01; //unknown 0x00 for some devices (mouse, keyboard), 0x02 for others (mouse Anywhere MX 2, presenter)
+            break;
     }
 
+
     self->tmp_tx_payload.data[13] = self->device_pairing_info.device_type;
-    if (g_logitacker_global_config.workmode == OPTION_LOGITACKER_WORKMODE_UNIFYING) {
-        self->tmp_tx_payload.data[14] = self->device_pairing_info.device_caps; // should have LOGITACKER_DEVICE_CAPS_UNIFYING_COMPATIBLE set and LOGITACKER_DEVICE_CAPS_LINK_ENCRYPTION unset
-    } else {
-        self->tmp_tx_payload.data[14] = 0xB3; // should have LOGITACKER_DEVICE_CAPS_UNIFYING_COMPATIBLE set and LOGITACKER_DEVICE_CAPS_LINK_ENCRYPTION unset
+
+
+    switch (g_logitacker_global_config.workmode) {
+        case OPTION_LOGITACKER_WORKMODE_UNIFYING:
+            self->tmp_tx_payload.data[14] = self->device_pairing_info.device_caps; // should have LOGITACKER_DEVICE_CAPS_UNIFYING_COMPATIBLE set and LOGITACKER_DEVICE_CAPS_LINK_ENCRYPTION unset
+            break;
+        case OPTION_LOGITACKER_WORKMODE_LIGHTSPEED:
+            self->tmp_tx_payload.data[14] = 0xB3; // should have LOGITACKER_DEVICE_CAPS_UNIFYING_COMPATIBLE set and LOGITACKER_DEVICE_CAPS_LINK_ENCRYPTION set
+            break;
+        case OPTION_LOGITACKER_WORKMODE_G700:
+            self->tmp_tx_payload.data[14] = 0x36; // should have LOGITACKER_DEVICE_CAPS_UNIFYING_COMPATIBLE set and LOGITACKER_DEVICE_CAPS_LINK_ENCRYPTION unset
+            break;
     }
+
 
     //15 is unknown