+++ logitacker/logitacker_radio.c
@@ -147,6 +147,7 @@ uint32_t logitacker_radio_init(nrf_esb_event_handler_t event_handler, radio_even
         case OPTION_LOGITACKER_WORKMODE_LIGHTSPEED:
             nrf_esb_update_channel_frequency_table_lightspeed();
             break;
+        case OPTION_LOGITACKER_WORKMODE_G700:
         case OPTION_LOGITACKER_WORKMODE_UNIFYING:
             nrf_esb_update_channel_frequency_table_unifying();
             break;